requirements.txt
chainlit==1.1.304
langchain==0.2.6
langchain-community==0.2.6
langchain-huggingface==0.0.3
langchain-groq==0.1.6
faiss-cpu==1.8.0
pypdf==4.2.0
python-dotenv==1.0.1
tiktoken
pydantic==2.9.2